+++ src/main/kotlin/org/jetbrains/intellij/platform/gradle/extensions/IntelliJPlatformDependenciesHelper.kt
@@ -17,6 +17,7 @@ import org.gradle.api.initialization.resolve.RulesMode
 import org.gradle.api.model.ObjectFactory
 import org.gradle.api.provider.Provider
 import org.gradle.api.provider.ProviderFactory
+import org.gradle.api.resources.ResourceHandler
 import org.gradle.internal.os.OperatingSystem
 import org.gradle.kotlin.dsl.*
 import org.jetbrains.intellij.platform.gradle.*
@@ -60,6 +61,7 @@ class IntelliJPlatformDependenciesHelper(
     private val layout: ProjectLayout,
     private val objects: ObjectFactory,
     private val providers: ProviderFactory,
+    private val resources: ResourceHandler,
     private val rootProjectDirectory: Path,
     private val metadataRulesModeProvider: Provider<RulesMode>,
 ) {
@@ -591,6 +593,8 @@ class IntelliJPlatformDependenciesHelper(
         providers.of(ProductReleasesValueSource::class.java) {
             parameters.jetbrainsIdesUrl = providers[GradleProperties.ProductsReleasesJetBrainsIdesUrl]
             parameters.androidStudioUrl = providers[GradleProperties.ProductsReleasesAndroidStudioUrl]
+            parameters.jetbrainsIdes = parameters.jetbrainsIdesUrl.map { resources.resolveUrl(it) }
+            parameters.androidStudio = parameters.androidStudioUrl.map { resources.resolveUrl(it) }
 
             parameters(configure)
         }