+++ src/main/kotlin/org/jetbrains/intellij/platform/gradle/plugins/project/IntelliJPlatformModulePlugin.kt
@@ -188,7 +188,7 @@ abstract class IntelliJPlatformModulePlugin : Plugin<Project> {
 
         // TODO: share with Base plugin?
         val dependenciesHelper = with(project) {
-            IntelliJPlatformDependenciesHelper(configurations, dependencies, layout, objects, providers, rootProjectPath, project.settings.dependencyResolutionManagement.rulesMode)
+            IntelliJPlatformDependenciesHelper(configurations, dependencies, layout, objects, providers, resources, rootProjectPath, project.settings.dependencyResolutionManagement.rulesMode)
         }
 
         IntelliJPlatformTestingExtension.register(project, dependenciesHelper, target = project)